+++ R/modelsummary.R
@@ -11,7 +11,7 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low', 'val
 #' @param output filename or object type (character string)
 #' \itemize{
 #'   \item Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.
-#'   \item Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
+#'   \item Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
 #'   \item Warning: the `output` argument \emph{cannot} be used when customizing tables with external packages. See the 'Details' section below.
 #' }
 #' @param fmt determines how to format numeric values
@@ -75,15 +75,17 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low', 'val
 #' the data.frame produced by `get_estimates(model)`. Examples:
 #' \itemize{
 #'   \item "estimate"
-#'   \item "\{estimate\} (\{std.error\}){stars}"
+#'   \item "\{estimate\} (\{std.error\})\{stars\}"
 #'   \item "\{estimate\} [\{conf.low\}, \{conf.high\}]"
 #' }
 #' @param align A character string of length equal to the number of columns in
 #' the table.  "lcr" means that the first column will be left-aligned, the 2nd
 #' column center-aligned, and the 3rd column right-aligned.
-#' @param ... all other arguments are passed to the `tidy` and `glance` methods
-#' used to extract estimates from the model. For example, this allows users to
-#' set `exponentiate=TRUE` to exponentiate logistic regression coefficients.
+#' @param ... all other arguments are passed through to the extractor and
+#' table-making functions. This allows users to specify additional options such
+#' as set `broom::tidy(exponentiate=TRUE)` to exponentiate logistic regression
+#' coefficients or `kableExtra::kbl(escape=FALSE)` to avoid escaping math
+#' characters in `kableExtra` tables.
 #' @return a regression table in a format determined by the `output` argument.
 #' @importFrom generics glance tidy
 #' @details 
@@ -306,7 +308,12 @@ modelsummary <- function(
 
     # coef_rename: before merge to collapse rows
     if (!is.null(coef_rename)) {
-      tmp$term <- replace_dict(tmp$term, coef_rename)
+      if (is.character(coef_rename)) {
+        dict <- coef_rename
+      } else if (is.function(coef_rename)) {
+        dict <- stats::setNames(coef_rename(tmp$term), tmp$term)
+      }
+      tmp$term <- replace_dict(tmp$term, dict)
     }
 
     # coef_map